move checkmarket job to manager
